try fix mingw warning in call to ioctlsocket, switch package build to ubuntu xenial base container so it is maintained

+++ src/rdwin32.h
@@ -269,7 +269,7 @@ typedef WSAPOLLFD rd_pollfd_t;
  * @returns 0 on success or -1 on failure (see rd_kafka_rd_socket_errno)
  */
 static RD_UNUSED int rd_fd_set_nonblocking (rd_socket_t fd) {
-        int on = 1;
+        u_long on = 1;
         if (ioctlsocket(fd, FIONBIO, &on) == SOCKET_ERROR)
                 return (int)WSAGetLastError();
         return 0;
